requirement.txt
pyyaml
torch
torchvision
torchaudio
matplotlib
tqdm
text2vec
pdfplumber
opencc-python-reimplemented